requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
Django==1.11.2
facebook-sdk==2.0.0
future==0.16.0
idna==2.6
kombu==4.1.0
nltk==3.2.5
numpy==1.13.3
oauthlib==2.0.6
pbr==3.1.1
psycopg2==2.7.3.2
python-twitter==3.3
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
stevedore==1.28.0
textblob==0.15.0
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
gunicorn==19.7.1
whitenoise==3.3.0
stripe==1.79.1

